+++ src/Laravel/ZaloServiceProvider.php
@@ -23,6 +23,8 @@ use FieldVn\Zalo\Laravel\Console\OaListCommand;
 use FieldVn\Zalo\Laravel\Console\OaTestCommand;
 use FieldVn\Zalo\Laravel\Console\RefreshTokensCommand;
 use FieldVn\Zalo\Laravel\Console\StatusCommand;
+use FieldVn\Zalo\Laravel\Console\ZbsSendCommand;
+use FieldVn\Zalo\Laravel\Console\ZbsTemplatesCommand;
 use FieldVn\Zalo\Laravel\Http\Middleware\Authorize;
 use FieldVn\Zalo\Laravel\Managers\ZaloManager;
 use FieldVn\Zalo\Laravel\Models\ZaloBot;
@@ -124,6 +126,8 @@ class ZaloServiceProvider extends ServiceProvider
                 BotSendCommand::class,
                 BotChatsCommand::class,
                 BotWebhookCommand::class,
+                ZbsTemplatesCommand::class,
+                ZbsSendCommand::class,
             ]);
         }
     }